Factor out the tests in `test_universe.py` to distinct files
* `test_universe_shape0d.py`:
  * `BersteinTB`, `BersteinPolynomialTC`, `BernsteinInterpolationTC`
  * `Point3dTB`, `Point3dFp32TC`, `Point3dFp64TC`
* `test_universe_shape1d.py`:
  * `Segment3dTB`, `Segment3dFp32TC`, `Segment3dFp64TC`
  * `Bezier3dTB`, `Bezier3dFp32TC`, `Segment3dFp64TC`
  * `SegmentPadTB`, `SegmentPadFp32TC`, `SegmentPadFp64TC`
  * `CurvePadTB`, `CurvePadFp32TC`, `CurvePadFp64TC`
* `test_universe_shape2d.py`
  * `Triangle3dTB`, `Triangle3dFp32TC`, `Triangle3dFp64TC`
  * `TrianglePadTB`, `TrianglePadFp32TC`, `TrianglePadFp64TC`

There is not unit tests for 3-D shapes yet.

The prompts used through clion claude plugin are:
- Split the tests in `test_universe.py` into `test_universe_shape0d.py`,
  `test_universe_shape1d.py`, `test_universe_shape2d.py`,
  `test_universe_shape3d.py` based on the wrapped C++ classes that are wrapped
  in `wrap_shape0d.cpp`, `wrap_shape1d.cpp`, `wrap_shape2d.cpp`,
  `wrap_shape3d.cpp` under `cpp/modmesh/universe/pymod/`.
